requirements.txt
# BlocIQ Production Check Dependencies

# Database connectivity (for schema checks)
psycopg2-binary>=2.9.0

# Core dependencies from BlocIQ_Onboarder
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-dateutil>=2.8.0
rapidfuzz>=3.0.0
PyYAML>=6.0.0
reportlab>=4.0.0
matplotlib>=3.5.0

# Supabase client (if needed for smoke tests)
supabase>=2.0.0

# Additional dependencies for extractors
lxml>=4.9.0
Pillow>=10.0.0
